# web/src/api/openapi.json
{
 "components": {
  "schemas": {
   "Candidate": {
    "additionalProperties": true,
    "properties": {
     "cell_id": {
      "title": "Cell Id",
      "type": "string"
     },
     "in_basin": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "In Basin"
     },
     "km_to_label": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Km To Label"
     },
     "known_share": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Known Share"
     },
     "label_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label Name"
     },
     "label_tier": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Label Tier"
     },
     "lat": {
      "title": "Lat",
      "type": "number"
     },
     "lon": {
      "title": "Lon",
      "type": "number"
     },
     "score": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Score"
     }
    },
    "required": [
     "cell_id",
     "lon",
     "lat"
    ],
    "title": "Candidate",
    "type": "object"
   },
   "CellConversations": {
    "properties": {
     "cell_id": {
      "title": "Cell Id",
      "type": "string"
     },
     "conversations": {
      "items": {
       "$ref": "#/components/schemas/ConversationSummary"
      },
      "title": "Conversations",
      "type": "array"
     }
    },
    "required": [
     "cell_id",
     "conversations"
    ],
    "title": "CellConversations",
    "type": "object"
   },
   "Cells": {
    "properties": {
     "cells": {
      "items": {
       "$ref": "#/components/schemas/Candidate"
      },
      "title": "Cells",
      "type": "array"
     }
    },
    "required": [
     "cells"
    ],
    "title": "Cells",
    "type": "object"
   },
   "ChatRequest": {
    "properties": {
     "cell_id": {
      "description": "looks like 0123_0045",
      "pattern": "^\\d{4}_\\d{4}$",
      "title": "Cell Id",
      "type": "string"
     },
     "conversation_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Conversation Id"
     },
     "question": {
      "maxLength": 4000,
      "minLength": 1,
      "title": "Question",
      "type": "string"
     }
    },
    "required": [
     "cell_id",
     "question"
    ],
    "title": "ChatRequest",
    "type": "object"
   },
   "ChatResponse": {
    "properties": {
     "cell_id": {
      "title": "Cell Id",
      "type": "string"
     },
     "conversation_id": {
      "title": "Conversation Id",
      "type": "string"
     },
     "cost_usd": {
      "title": "Cost Usd",
      "type": "number"
     },
     "turn": {
      "$ref": "#/components/schemas/Turn"
     },
     "values": {
      "additionalProperties": {
       "$ref": "#/components/schemas/Val"
      },
      "title": "Values",
      "type": "object"
     }
    },
    "required": [
     "conversation_id",
     "cell_id",
     "turn",
     "values",
     "cost_usd"
    ],
    "title": "ChatResponse",
    "type": "object"
   },
   "Claim": {
    "additionalProperties": true,
    "properties": {
     "claim_no": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Claim No"
     },
     "text": {
      "title": "Text",
      "type": "string"
     },
     "value_ids": {
      "items": {
       "type": "string"
      },
      "title": "Value Ids",
      "type": "array"
     }
    },
    "required": [
     "text"
    ],
    "title": "Claim",
    "type": "object"
   },
   "ConversationRecord": {
    "properties": {
     "backend": {
      "title": "Backend",
      "type": "string"
     },
     "cell_id": {
      "title": "Cell Id",
      "type": "string"
     },
     "conversation_id": {
      "title": "Conversation Id",
      "type": "string"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "model": {
      "title": "Model",
      "type": "string"
     },
     "turns": {
      "items": {
       "$ref": "#/components/schemas/StoredTurn"
      },
      "title": "Turns",
      "type": "array"
     }
    },
    "required": [
     "conversation_id",
     "cell_id",
     "model",
     "backend",
     "created_at",
     "turns"
    ],
    "title": "ConversationRecord",
    "type": "object"
   },
   "ConversationSummary": {
    "properties": {
     "backend": {
      "title": "Backend",
      "type": "string"
     },
     "conversation_id": {
      "title": "Conversation Id",
      "type": "string"
     },
     "cost_usd": {
      "title": "Cost Usd",
      "type": "number"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "model": {
      "title": "Model",
      "type": "string"
     },
     "turns": {
      "title": "Turns",
      "type": "integer"
     }
    },
    "required": [
     "conversation_id",
     "model",
     "backend",
     "created_at",
     "turns",
     "cost_usd"
    ],
    "title": "ConversationSummary",
    "type": "object"
   },
   "Evidence": {
    "properties": {
     "cell_id": {
      "title": "Cell Id",
      "type": "string"
     },
     "in_basin": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "In Basin"
     },
     "lat": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lat"
     },
     "lon": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Lon"
     },
     "memos": {
      "items": {
       "$ref": "#/components/schemas/Memo"
      },
      "title": "Memos",
      "type": "array"
     },
     "parts": {
      "additionalProperties": {
       "$ref": "#/components/schemas/ToolResultOut"
      },
      "title": "Parts",
      "type": "object"
     },
     "values": {
      "additionalProperties": {
       "$ref": "#/components/schemas/Val"
      },
      "title": "Values",
      "type": "object"
     }
    },
    "required": [
     "cell_id",
     "parts",
     "values",
     "memos"
    ],
    "title": "Evidence",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
   "Health": {
    "properties": {
     "backend": {
      "title": "Backend",
      "type": "string"
     },
     "effort": {
      "title": "Effort",
      "type": "string"
     },
     "model": {
      "title": "Model",
      "type": "string"
     },
     "ok": {
      "title": "Ok",
      "type": "boolean"
     }
    },
    "required": [
     "ok",
     "model",
     "effort",
     "backend"
    ],
    "title": "Health",
    "type": "object"
   },
   "Memo": {
    "properties": {
     "claims": {
      "items": {
       "$ref": "#/components/schemas/Claim"
      },
      "title": "Claims",
      "type": "array"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "memo_id": {
      "title": "Memo Id",
      "type": "string"
     },
     "published": {
      "title": "Published",
      "type": "boolean"
     },
     "role": {
      "title": "Role",
      "type": "string"
     },
     "verdict": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Verdict"
     }
    },
    "required": [
     "memo_id",
     "role",
     "published",
     "created_at",
     "claims"
    ],
    "title": "Memo",
    "type": "object"
   },
   "StoredTurn": {
    "properties": {
     "claims": {
      "items": {
       "$ref": "#/components/schemas/Claim"
      },
      "title": "Claims",
      "type": "array"
     },
     "cost_usd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost Usd"
     },
     "created_at": {
      "title": "Created At",
      "type": "string"
     },
     "duration_s": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Duration S"
     },
     "problems": {
      "items": {
       "type": "string"
      },
      "title": "Problems",
      "type": "array"
     },
     "published": {
      "title": "Published",
      "type": "boolean"
     },
     "question": {
      "title": "Question",
      "type": "string"
     },
     "step": {
      "title": "Step",
      "type": "integer"
     },
     "text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Text"
     },
     "tool_calls": {
      "items": {
       "$ref": "#/components/schemas/ToolCall"
      },
      "title": "Tool Calls",
      "type": "array"
     },
     "values": {
      "additionalProperties": {
       "$ref": "#/components/schemas/Val"
      },
      "title": "Values",
      "type": "object"
     }
    },
    "required": [
     "step",
     "question",
     "published",
     "problems",
     "claims",
     "tool_calls",
     "values",
     "created_at"
    ],
    "title": "StoredTurn",
    "type": "object"
   },
   "ToolCall": {
    "additionalProperties": true,
    "properties": {
     "args": {
      "additionalProperties": true,
      "title": "Args",
      "type": "object"
     },
     "file": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "File"
     },
     "rows": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Rows"
     },
     "tool": {
      "title": "Tool",
      "type": "string"
     },
     "values": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Values"
     }
    },
    "required": [
     "tool",
     "args"
    ],
    "title": "ToolCall",
    "type": "object"
   },
   "ToolResultOut": {
    "properties": {
     "args": {
      "additionalProperties": true,
      "title": "Args",
      "type": "object"
     },
     "note": {
      "default": "",
      "title": "Note",
      "type": "string"
     },
     "rows": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Rows",
      "type": "array"
     },
     "tool": {
      "title": "Tool",
      "type": "string"
     },
     "values": {
      "additionalProperties": {
       "$ref": "#/components/schemas/Val"
      },
      "title": "Values",
      "type": "object"
     }
    },
    "required": [
     "tool",
     "args",
     "rows",
     "values"
    ],
    "title": "ToolResultOut",
    "type": "object"
   },
   "Turn": {
    "additionalProperties": true,
    "properties": {
     "asked_at": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Asked At"
     },
     "cannot_answer": {
      "default": false,
      "title": "Cannot Answer",
      "type": "boolean"
     },
     "caveats": {
      "items": {
       "type": "string"
      },
      "title": "Caveats",
      "type": "array"
     },
     "claims": {
      "items": {
       "$ref": "#/components/schemas/Claim"
      },
      "title": "Claims",
      "type": "array"
     },
     "cost_usd": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cost Usd"
     },
     "problems": {
      "items": {
       "type": "string"
      },
      "title": "Problems",
      "type": "array"
     },
     "published": {
      "title": "Published",
      "type": "boolean"
     },
     "question": {
      "title": "Question",
      "type": "string"
     },
     "text": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Text"
     },
     "tools_used": {
      "items": {
       "type": "string"
      },
      "title": "Tools Used",
      "type": "array"
     },
     "usage_limited": {
      "default": false,
      "title": "Usage Limited",
      "type": "boolean"
     }
    },
    "required": [
     "question",
     "published"
    ],
    "title": "Turn",
    "type": "object"
   },
   "Val": {
    "additionalProperties": true,
    "properties": {
     "as_printed": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "As Printed"
     },
     "fmt": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Fmt"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "kind": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Kind"
     },
     "note": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Note"
     },
     "unit": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Unit"
     },
     "unit_as_printed": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Unit As Printed"
     },
     "value": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "integer"
       },
       {
        "type": "string"
       },
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Value"
     }
    },
    "required": [
     "id"
    ],
    "title": "Val",
    "type": "object"
   },
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  }
 },
 "info": {
  "title": "AI Uranium Explorer service",
  "version": "0.2.0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/cell/{cell_id}": {
   "get": {
    "operationId": "cell_api_cell__cell_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "cell_id",
      "required": true,
      "schema": {
       "title": "Cell Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Evidence"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Cell"
   }
  },
  "/api/cell/{cell_id}/conversations": {
   "get": {
    "operationId": "cell_conversations_api_cell__cell_id__conversations_get",
    "parameters": [
     {
      "in": "path",
      "name": "cell_id",
      "required": true,
      "schema": {
       "title": "Cell Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CellConversations"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Cell Conversations"
   }
  },
  "/api/cells": {
   "get": {
    "operationId": "cells_api_cells_get",
    "parameters": [
     {
      "in": "query",
      "name": "limit",
      "required": false,
      "schema": {
       "default": 40,
       "maximum": 200,
       "minimum": 1,
       "title": "Limit",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "model",
      "required": false,
      "schema": {
       "default": "criteria",
       "title": "Model",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Cells"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Cells"
   }
  },
  "/api/chat": {
   "post": {
    "operationId": "chat_api_chat_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChatRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ChatResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Chat"
   }
  },
  "/api/chat/stream": {
   "post": {
    "operationId": "chat_stream_api_chat_stream_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ChatRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Chat Stream"
   }
  },
  "/api/conversation/{conversation_id}": {
   "get": {
    "operationId": "conversation_api_conversation__conversation_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "conversation_id",
      "required": true,
      "schema": {
       "title": "Conversation Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ConversationRecord"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Conversation"
   }
  },
  "/api/health": {
   "get": {
    "operationId": "health_api_health_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Health"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "Health"
   }
  }
 }
}